requirements.txt
#appdirs==1.4.4
#attrs==19.3.0
#bcrypt==3.1.7
#beautifulsoup4==4.8.2
#blinker==1.4
#cached-property==1.5.1
#certifi==2019.11.28
#cfgv==3.2.0
#chardet==3.0.4
#Click==7.0
#colorama==0.4.3
#configobj==5.0.6
#cryptography==2.8
#cycler==0.10.0
#dbus-python==1.2.16
#distlib==0.3.1
#distro==1.4.0
#docker==4.1.0
#docker-compose==1.25.0
#dockerpty==0.4.1
#docopt==0.6.2
#entrypoints==0.3
#fasteners==0.14.1
#filelock==3.0.12
#future==0.18.2
#html5lib==1.0.1
#httpie==1.0.3
#httplib2==0.14.0
#identify==2.2.4
#idna==2.8
#importlib-metadata==1.5.0
#jsonschema==3.2.0
#keyring==18.0.1
#kiwisolver==1.3.1
#launchpadlib==1.10.13
#lazr.restfulclient==0.14.2
#lazr.uri==1.0.3
#libevdev==0.5
#lockfile==0.12.2
#lxml==4.5.0
#macaroonbakery==1.3.1
#Mako==1.1.0
#MarkupSafe==1.1.0
#monotonic==1.5
#more-itertools==4.2.0
#mpmath==1.2.1
#netifaces==0.10.4
#nodeenv==1.6.0
#oauthlib==3.1.0
#olefile==0.46
#paramiko==2.6.0
#pexpect==4.6.0
Pillow==7.0.0
#protobuf==3.6.1
#psutil==5.5.1
#pycairo==1.16.2
#pycups==1.9.73
#Pygments==2.3.1
PyJWT==1.7.1
#pymacaroons==0.13.0
#PyNaCl==1.3.0
#pyparsing==2.4.7
#pyRFC3339==1.1
#pyrsistent==0.15.5
#python-dateutil==2.7.3
#python-xlib==0.23
#pytz==2019.3
#pyudev==0.21.0
#pyxdg==0.26
#reportlab==3.5.34
#requests==2.22.0
#requests-unixsocket==0.2.0
#SecretStorage==2.3.1
#setproctitle==1.1.10
#simplejson==3.16.0
#six==1.14.0
#soupsieve==1.9.5
#sympy==1.8
#texttable==1.6.2
#toml==0.10.2
#SecretStorage==2.3.1
#setproctitle==1.1.10
#simplejson==3.16.0
#six==1.14.0
#soupsieve==1.9.5
#sympy==1.8
#testresources==2.0.1
#texttable==1.6.2
#toml==0.10.2
#urllib3==1.25.8
#virtualenv==20.4.6
#wadllib==1.3.3
#webencodings==0.5.1
#websocket-client==0.53.0
#zipp==1.0.0
bcrypt==3.1.7
# certifi==2019.11.28
# chardet==3.0.4
# Click==7.0
# colorama==0.4.3
# coreapi==2.3.3
# coreschema==0.0.4
# defusedxml==0.7.1
Django==3.2.3
django-cors-headers==3.7.0
django-settings-json==1.0.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
# djangorestframework-simplejwt==4.4.0
djoser==2.1.0
entrypoints==0.3
# fasteners==0.14.1
future==0.18.2
# httpie==1.0.3
httplib2==0.14.0
# idna==2.8
# itypes==1.2.0
# Jinja2==3.0.1
# keyring==18.0.1
# launchpadlib==1.10.13
# lazr.restfulclient==0.14.2
# lazr.uri==1.0.3
# lockfile==0.12.2
# macaroonbakery==1.3.1
# Mako==1.1.0
# MarkupSafe==2.0.1
# monotonic==1.5
# netifaces==0.10.4
# oauthlib==3.1.0
# olefile==0.46
# paramiko==2.6.0
# pexpect==4.6.0
# Pillow==7.0.0
# protobuf==3.6.1
# psutil==5.5.1
psycopg2-binary==2.8.6
# pycairo==1.16.2
# Pygments==2.3.1
PyJWT==1.7.1
# pymacaroons==0.13.0
# PyNaCl==1.3.0
# pyRFC3339==1.1
# python-dateutil==2.7.3
# python3-openid==3.2.0
# pytz==2019.3
# pyxdg==0.26
# reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
# SecretStorage==2.3.1
simplejson==3.16.0
social-auth-app-django==4.0.0
# social-auth-core==4.1.0
# sqlparse==0.4.1
# uritemplate==3.0.1
# urllib3==1.25.8
# wadllib==1.3.3
django-extensions
celery~=3.1.26.post2